+++ internal/query/session.go
@@ -4,6 +4,7 @@ import (
 	"context"
 
 	"github.com/ydb-platform/ydb-go-genproto/Ydb_Query_V1"
+	"github.com/ydb-platform/ydb-go-genproto/protos/Ydb"
 
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/query/options"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/query/result"
@@ -23,7 +24,7 @@ type (
 
 		client Ydb_Query_V1.QueryServiceClient
 		trace  *trace.Query
-		laztTx bool
+		lazyTx bool
 	}
 )
 
@@ -38,6 +39,8 @@ func (s *Session) QueryResultSet(
 
 	r, err := execute(ctx, s.ID(), s.client, q, options.ExecuteSettings(opts...), withTrace(s.trace))
 	if err != nil {
+		s.setStatusFromError(err)
+
 		return nil, xerrors.WithStackTrace(err)
 	}
 
@@ -54,6 +57,8 @@ func (s *Session) queryRow(
 ) (row query.Row, finalErr error) {
 	r, err := execute(ctx, s.ID(), s.client, q, settings, resultOpts...)
 	if err != nil {
+		s.setStatusFromError(err)
+
 		return nil, xerrors.WithStackTrace(err)
 	}
 
@@ -111,7 +116,7 @@ func (s *Session) Begin(
 		}
 	}()
 
-	if s.laztTx {
+	if s.lazyTx {
 		return &Transaction{
 			s:          s,
 			txSettings: txSettings,
@@ -120,6 +125,8 @@ func (s *Session) Begin(
 
 	txID, err := begin(ctx, s.client, s.ID(), txSettings)
 	if err != nil {
+		s.setStatusFromError(err)
+
 		return nil, xerrors.WithStackTrace(err)
 	}
 
@@ -140,6 +147,8 @@ func (s *Session) Exec(
 
 	r, err := execute(ctx, s.ID(), s.client, q, options.ExecuteSettings(opts...), withTrace(s.trace))
 	if err != nil {
+		s.setStatusFromError(err)
+
 		return xerrors.WithStackTrace(err)
 	}
 
@@ -162,8 +171,21 @@ func (s *Session) Query(
 
 	r, err := execute(ctx, s.ID(), s.client, q, options.ExecuteSettings(opts...), withTrace(s.trace))
 	if err != nil {
+		s.setStatusFromError(err)
+
 		return nil, xerrors.WithStackTrace(err)
 	}
 
 	return r, nil
+}
+
+func (s *Session) setStatusFromError(err error) {
+	switch {
+	case xerrors.IsTransportError(err):
+		s.SetStatus(session.StatusError)
+	case xerrors.IsOperationError(err, Ydb.StatusIds_SESSION_BUSY, Ydb.StatusIds_BAD_SESSION):
+		s.SetStatus(session.StatusError)
+	case xerrors.IsOperationError(err, Ydb.StatusIds_BAD_SESSION):
+		s.SetStatus(session.StatusClosed)
+	}
 }