+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Fixed drop session from pool unnecessary in query service 
+
 ## v3.96.0
 * Supported of list, set and struct for unmarshall using `sugar.Unmarshall...`